requirements.txt
# PyWavelets
# opencv-python
# seaborn
# Gunicorn
# selenium
# seaborn
# scikit-learn
# requests
# pytz
# pyttsx3
# python-dotenv
# pandas
# matplotlib
# Flask
# Flask-Cors 
# google-ai-generativelanguage
# google-auth
# google-generativeai
# Pillow
# translators
# numpy
# json5
# joblib

PyWavelets==1.5.0
opencv-python==4.7.0.72
seaborn==0.13.0
gunicorn==21.2.0
selenium==4.8.3
scikit-learn==1.3.1
requests==2.31.0
pytz==2022.7
pyttsx3==2.90
python-dotenv==1.0.0
pandas==2.1.1
matplotlib==3.8.0
Flask==3.0.0
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-auth==2.17.3
google-generativeai==0.3.2
Pillow==9.5.0
translators==5.8.9
numpy==1.26.4
json5==0.9.14
joblib==1.3.2
opencv-python-headless==4.7.0.72